third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-needkey-expected.txt

This tests that the 'needkey' event is generated.

RUN(video.src = "../content/test-encrypted.webm")
EVENT(needkey)
EXPECTED (event.target == '[object HTMLVideoElement]') OK
EXPECTED (event instanceof window.MediaKeyNeededEvent == 'true') OK
EXPECTED (event.type == 'needkey') OK
EXPECTED (event.contentType == 'video/webm') OK
EXPECTED (event.initData.length == '16') OK
EXPECTED (event.initData[0] == '48') OK
EXPECTED (event.initData[1] == '49') OK
EXPECTED (event.initData[2] == '50') OK
EXPECTED (event.initData[3] == '51') OK
EXPECTED (event.initData[4] == '52') OK
EXPECTED (event.initData[5] == '53') OK
EXPECTED (event.initData[6] == '54') OK
EXPECTED (event.initData[7] == '55') OK
EXPECTED (event.initData[8] == '56') OK
EXPECTED (event.initData[9] == '57') OK
EXPECTED (event.initData[10] == '48') OK
EXPECTED (event.initData[11] == '49') OK
EXPECTED (event.initData[12] == '50') OK
EXPECTED (event.initData[13] == '51') OK
EXPECTED (event.initData[14] == '52') OK
EXPECTED (event.initData[15] == '53') OK
EVENT(needkey)
EXPECTED (event.target == '[object HTMLVideoElement]') OK
EXPECTED (event instanceof window.MediaKeyNeededEvent == 'true') OK
EXPECTED (event.type == 'needkey') OK
EXPECTED (event.contentType == 'video/webm') OK
EXPECTED (event.initData.length == '16') OK
EXPECTED (event.initData[0] == '48') OK
EXPECTED (event.initData[1] == '49') OK
EXPECTED (event.initData[2] == '50') OK
EXPECTED (event.initData[3] == '51') OK
EXPECTED (event.initData[4] == '52') OK
EXPECTED (event.initData[5] == '53') OK
EXPECTED (event.initData[6] == '54') OK
EXPECTED (event.initData[7] == '55') OK
EXPECTED (event.initData[8] == '56') OK
EXPECTED (event.initData[9] == '57') OK
EXPECTED (event.initData[10] == '48') OK
EXPECTED (event.initData[11] == '49') OK
EXPECTED (event.initData[12] == '50') OK
EXPECTED (event.initData[13] == '51') OK
EXPECTED (event.initData[14] == '52') OK
EXPECTED (event.initData[15] == '53') OK
END OF TEST
